# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OmniPlot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/preview/stream": {
      "get": {
        "description": "Open a Server-Sent Events stream emitting progress events.\n\nWhen ``file_id`` is set, runs the real pipeline against the\nmatching library entry; otherwise falls back to the synthetic\nemitter (useful for tests and frontend smoke).\n\nRaises:\n    HTTPException: 404 when ``file_id`` is unknown or its original\n        bytes are missing on disk \u2014 validated *before* the streaming\n        response starts so the client gets a real 404 status instead\n        of an error frame behind a 200.",
        "operationId": "preview_stream_preview_stream_get",
        "parameters": [
          {
            "in": "query",
            "name": "layer_count",
            "required": false,
            "schema": {
              "default": 4,
              "maximum": 64,
              "minimum": 1,
              "title": "Layer Count",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "file_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "File Id"
            }
          },
          {
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Stream"
      }
    },
...
  }
}